fix(pointers): [iOS] Fix ListView scrolling with manipulation
