+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/NativeListViewBase.iOS.cs
@@ -11,6 +11,7 @@ using Uno.Disposables;
 using Windows.UI.Core;
 using Uno.UI.Controls;
 using System.ComponentModel;
+using Windows.UI.Input;
 using Uno.Extensions.Specialized;
 using Windows.UI.Xaml.Controls.Primitives;
 using Uno.Logging;
@@ -81,34 +82,21 @@ namespace Windows.UI.Xaml.Controls
 		#region Properties
 		new internal ListViewBaseSource Source
 		{
-			get { return base.Source as ListViewBaseSource; }
-			set
-			{
-				base.Source = value;
-			}
+			get => base.Source as ListViewBaseSource;
+			set => base.Source = value;
 		}
 
 		public Style ItemContainerStyle => XamlParent?.ItemContainerStyle;
 
-		public DataTemplate HeaderTemplate
-		{
-			get { return XamlParent?.HeaderTemplate; }
-		}
+		public DataTemplate HeaderTemplate => XamlParent?.HeaderTemplate;
 
-
-		public DataTemplate FooterTemplate
-		{
-			get { return XamlParent?.FooterTemplate; }
-		}
-
+		public DataTemplate FooterTemplate => XamlParent?.FooterTemplate;
 
 		public DataTemplateSelector ItemTemplateSelector => XamlParent?.ItemTemplateSelector;
 
 		internal bool NeedsReloadData => _needsReloadData;
 
 		internal CGPoint UpperScrollLimit { get { return (CGPoint)(ContentSize - Frame.Size); } }
-
-		internal UIElement.TouchesManager TouchesManager { get; /* readonly in int */ private set; }
 		#endregion
 
 		public GroupStyle GroupStyle => XamlParent?.GroupStyle.FirstOrDefault();
@@ -118,15 +106,10 @@ namespace Windows.UI.Xaml.Controls
 		internal IList<object> SelectedItems => XamlParent?.SelectedItems;
 
 		public ListViewSelectionMode SelectionMode => XamlParent?.SelectionMode ?? ListViewSelectionMode.None;
-		public object Header
-		{
-			get { return XamlParent?.ResolveHeaderContext(); }
-		}
 
-		public object Footer
-		{
-			get { return XamlParent?.ResolveFooterContext(); }
-		}
+		public object Header => XamlParent?.ResolveHeaderContext();
+
+		public object Footer => XamlParent?.ResolveFooterContext();
 
 		/// <summary>
 		/// Get all currently visible supplementary views.
@@ -178,8 +161,7 @@ namespace Windows.UI.Xaml.Controls
 			RegisterClassForSupplementaryView(internalContainerType, ListViewFooterElementKindNS, ListViewFooterReuseIdentifier);
 			RegisterClassForSupplementaryView(internalContainerType, ListViewSectionHeaderElementKindNS, ListViewSectionHeaderReuseIdentifier);
 
-			DelaysContentTouches = true;
-			TouchesManager = UIElement.TouchesManager.GetOrCreate(this);
+			DelaysContentTouches = true; // cf. TouchesManager which can alter this!
 
 			ShowsHorizontalScrollIndicator = true;
 			ShowsVerticalScrollIndicator = true;
@@ -498,6 +480,8 @@ namespace Windows.UI.Xaml.Controls
 			}
 		}
 
+		internal Orientation ScrollOrientation => (CollectionViewLayout as VirtualizingPanelLayout)?.ScrollOrientation ?? Orientation.Vertical;
+
 		public ScrollBarVisibility HorizontalScrollBarVisibility
 		{
 			get
@@ -620,7 +604,7 @@ namespace Windows.UI.Xaml.Controls
 				return UICollectionViewScrollPosition.None;
 			}
 
-			var scrollDirection = (CollectionViewLayout as VirtualizingPanelLayout)?.ScrollOrientation ?? Orientation.Vertical;
+			var scrollDirection = ScrollOrientation;
 			var snapPointsAlignment = (CollectionViewLayout as VirtualizingPanelLayout)?.SnapPointsAlignment;
 
 			switch (scrollDirection)
@@ -666,7 +650,7 @@ namespace Windows.UI.Xaml.Controls
 				this.Log().Warn("ScrollIntoViewAlignment.Default is not implemented");
 			}
 
-			var scrollDirection = (CollectionViewLayout as VirtualizingPanelLayout)?.ScrollOrientation ?? Orientation.Vertical;
+			var scrollDirection = ScrollOrientation;
 			switch (scrollDirection)
 			{
 				case Orientation.Horizontal:
@@ -769,5 +753,38 @@ namespace Windows.UI.Xaml.Controls
 			get { return NativeLayout.Padding; }
 			set { NativeLayout.Padding = value; }
 		}
+
+		#region Touches
+		private TouchesManager _touchesManager;
+
+		internal TouchesManager TouchesManager => _touchesManager ??= new NativeListViewBaseTouchesManager(this);
+
+		private class NativeListViewBaseTouchesManager : TouchesManager
+		{
+			private readonly NativeListViewBase _listView;
+
+			public NativeListViewBaseTouchesManager(NativeListViewBase listView)
+			{
+				_listView = listView;
+			}
+
+			/// <inheritdoc />
+			protected override bool CanConflict(GestureRecognizer.Manipulation manipulation)
+				=> manipulation.IsDragManipulation || _listView.ScrollOrientation switch
+				{
+					Orientation.Horizontal => manipulation.IsTranslateXEnabled,
+					Orientation.Vertical => manipulation.IsTranslateYEnabled,
+					_ => manipulation.IsTranslateXEnabled || manipulation.IsTranslateYEnabled
+				};
+
+			/// <inheritdoc />
+			protected override void SetCanDelay(bool canDelay)
+				=> _listView.DelaysContentTouches = canDelay;
+
+			/// <inheritdoc />
+			protected override void SetCanCancel(bool canCancel)
+				=> _listView.CanCancelContentTouches = canCancel;
+		}
+		#endregion
 	}
 }
