+++ src/Uno.UI/UI/Xaml/Controls/ScrollContentPresenter/NativeScrollContentPresenter.iOS.cs
@@ -57,7 +57,6 @@ namespace Windows.UI.Xaml.Controls
 
 		public NativeScrollContentPresenter()
 		{
-			TouchesManager = new ScrollContentPresenterManipulationManager(this);
 			Scrolled += OnScrolled;
 			ViewForZoomingInScrollView = _ => Content as UIView;
 			DidZoom += OnZoom;
@@ -560,11 +559,12 @@ namespace Windows.UI.Xaml.Controls
 		 * On the UIElement this is defined by the ManipulationMode.
 		 */
 
-		internal UIElement.TouchesManager TouchesManager { get; }
+		private TouchesManager _touchesManager;
+		internal TouchesManager TouchesManager => _touchesManager ??= new NativeScrollContentPresenterManipulationManager(this);
 
 		private void UpdateDelayedTouches()
 		{
-			if (TouchesManager.Listeners == 0)
+			if ((_touchesManager?.Listeners ?? 0) == 0)
 			{
 				// This prevents unnecessary touch delays (which affects the pressed visual states of buttons) when user can't scroll.
 				var canScrollVertically = VerticalScrollBarVisibility != ScrollBarVisibility.Disabled && ContentSize.Height > Frame.Height;
@@ -577,11 +577,11 @@ namespace Windows.UI.Xaml.Controls
 			}
 		}
 
-		private class ScrollContentPresenterManipulationManager : UIElement.TouchesManager
+		private class NativeScrollContentPresenterManipulationManager : TouchesManager
 		{
 			private readonly NativeScrollContentPresenter _scrollPresenter;
 
-			public ScrollContentPresenterManipulationManager(NativeScrollContentPresenter scrollPresenter)
+			public NativeScrollContentPresenterManipulationManager(NativeScrollContentPresenter scrollPresenter)
 			{
 				_scrollPresenter = scrollPresenter;
 			}
@@ -589,8 +589,8 @@ namespace Windows.UI.Xaml.Controls
 			/// <inheritdoc />
 			protected override bool CanConflict(GestureRecognizer.Manipulation manipulation)
 				=> _scrollPresenter.CanHorizontallyScroll && manipulation.IsTranslateXEnabled
-				|| _scrollPresenter.CanVerticallyScroll && manipulation.IsTranslateYEnabled
-				|| manipulation.IsDragManipulation; // This will actually always be false when CanConflict is being invoked in current setup.
+					|| _scrollPresenter.CanVerticallyScroll && manipulation.IsTranslateYEnabled
+					|| manipulation.IsDragManipulation; // This will actually always be false when CanConflict is being invoked in current setup.
 
 			/// <inheritdoc />
 			protected override void SetCanDelay(bool canDelay)
@@ -600,6 +600,6 @@ namespace Windows.UI.Xaml.Controls
 			protected override void SetCanCancel(bool canCancel)
 				=> _scrollPresenter.CanCancelContentTouches = canCancel;
 		}
-	#endregion
+		#endregion
 	}
 }
